requirements.txt
attrs==22.1.0
Automat==22.10.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
constantly==15.1.0
contourpy==1.0.6
cryptography==38.0.3
cssselect==1.2.0
cycler==0.11.0
filelock==3.8.0
fonttools==4.38.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
kiwisolver==1.4.4
lxml==4.9.1
matplotlib==3.6.2
numpy==1.23.5
packaging==21.3
pandas==1.5.2
parsel==1.7.0
Pillow==9.3.0
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.6
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
seaborn==0.12.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.2.post1
text-unidecode==1.3
tldextract==3.4.0
Twisted==22.10.0
typing_extensions==4.4.0
urllib3==1.26.12
w3lib==2.0.1
zope.interface==5.5.1
